requirements.txt
gunicorn
astroid
boto3
botocore
coverage
Django
django-bootstrap-form
django-factory-boy
django-s3direct
django-test-without-migrations
docutils
factory-boy
Faker
flake8
isort
jmespath
lazy-object-proxy
mccabe
pep8
psycopg2
pycodestyle
pyflakes
pylint
python-dateutil
python-decouple
pytz
s3transfer
selenium
six
whitenoise
wrapt
django-imagekit
Pillow
django-registration-redux
